fixed bug where fusion furnace was eating lava buckets but not being fueled; changed burnTime variables to ints (was an overflow issue). BREAKING CHANGE

+++ src/main/java/mod/alexndr/fusion/api/content/AbstractAlloyFurnaceContainer.java
@@ -33,8 +33,8 @@ public abstract class AbstractAlloyFurnaceContainer<T extends AbstractAlloyFurna
         // Add tracking for data (Syncs to client/updates value when it changes)
         this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.smeltTimeLeft, v -> tileEntity.smeltTimeLeft = (short) v));
         this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.maxSmeltTime, v -> tileEntity.maxSmeltTime = (short) v));
-        this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.fuelBurnTimeLeft, v -> tileEntity.fuelBurnTimeLeft = (short) v));
-        this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.maxFuelBurnTime, v -> tileEntity.maxFuelBurnTime = (short) v));
+        this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.fuelBurnTimeLeft, v -> tileEntity.fuelBurnTimeLeft = (int) v));
+        this.trackInt(new FunctionalIntReferenceHolder(() -> tileEntity.maxFuelBurnTime, v -> tileEntity.maxFuelBurnTime = (int) v));
 
         // Add all the slots for the tileEntity's inventory and the playerInventory to this container
